Added more security checks

+++ skieasy_app/views.py
@@ -20,10 +20,6 @@ def __profile_check(action_function):
     def my_wrapper_function(request, *args, **kwargs):
         try:
             user = User.objects.get(id=request.user.id)
-            print("USER: ")
-            print(user)
-            print("PROFILE: ")
-            print(Profile.objects.filter(user=user))
             if (Profile.objects.filter(user=user).exists()):
                 return action_function(request, *args, **kwargs)
             else:
@@ -45,6 +41,7 @@ class HomeView(LoginRequiredMixin, FilterView):
     paginate_by = 12
     ordering = ['profile_id']
 
+    @__profile_check
     def get_queryset(self):
         queryset = super().get_queryset()
         queryset = (
@@ -155,6 +152,8 @@ def register(request):
 @__profile_check
 def equipment_details(request, id):
     e = Equipment.objects.get(id=id)
+    if (e.profile_id.user != request.user):
+        return redirect(display_equipment)
     context = {}
     equip = {
             "id": e.id,
@@ -333,6 +332,8 @@ def create_listing(request, id):
 @__profile_check
 def update_equipment(request, id):
     equip = Equipment.objects.get(id=id)
+    if (equip.profile_id.user != request.user):
+        return redirect(display_equipment)
     context = {}
     context['equip_id'] = id
     if (request.method == 'GET'):
@@ -370,6 +371,8 @@ def update_equipment(request, id):
 def update_listing(request, id):
     context = {}
     li = EquipmentListing.objects.get(id=id)
+    if (li.profile_id.user != request.user):
+        return redirect(display_equipment)
     context['title'] = li.equipment_id.title
     context['list_id'] = id
     context['form'] = EquipmentListingForm()
@@ -416,6 +419,8 @@ def rent_listing(request, id):
 @__profile_check
 def delete_equipment(request, id):
     equipment = Equipment.objects.filter(id=id)
+    if (equipment.profile_id.user != request.user):
+        return redirect(display_equipment)
     equipment_listings = EquipmentListing.objects.filter(equipment_id=id)
     equipment_listings.delete()
     equipment.delete()
